requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg==3.1.16
django-cors-headers==4.3.1
pytest==7.4.4
pytest-django==4.7.0
djangorestframework-jwt-4==1.12.0
six==1.16.0
pyyaml==6.0.1
uritemplate==4.1.1
drf-spectacular==0.27.1
